requirements.txt
# FastAPI y servidor web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# HTTP y comunicación
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.0
python-multipart>=0.0.6

# Templates y frontend
jinja2>=3.1.0
python-dotenv>=1.0.0

# Playwright - Navegación real
playwright>=1.40.0

# APIs externas
openai>=1.3.0
httpx>=0.25.0

# Base de datos
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilidades
python-dateutil>=2.8.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0
httpx>=0.25.0

# Desarrollo
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
